kv: narrow the ATTN_PACK rope tail to bf16 (712 -> 584 B/row)
The packed comp-KV row kept its 64-wide rope tail in f32 while the nope
dims rode e4m3, so 256 of 712 bytes carried precision the model never had:
DSv4-Flash runs a BF16 residual. Narrowing the tail takes the row to 584 B,
an 18.0% cut in comp-KV bytes on the rows that dominate the attention tile.

That lands the row BYTE-IDENTICAL to vLLM's fp8_ds_mla DSv4 cache line
(448 e4m3 + 7 ue8m0-per-64 + 1 pad + 64 bf16 = 584). Every other field
already matched: same nope width, same group size, same scale dtype, same
7+1 padding, same +127 bias. The rope dtype was the sole divergence --
mainline moved DSv3's per-128 f32 scales to per-64 ue8m0 for DSv4, i.e.
onto what this cache already did. Matching the layout buys a cross-engine
KV diff against a stock vLLM run, a sharper fidelity check than comparing
logits.

The value-preserving invariant is kept, and that is the load-bearing
detail. attn_pack_store_kernel already wrote the fp8 roundtrip back into
the f32 staging buffer so the packed row and the shadow hold the same
numbers; it now does the same for rope, roundtripping the tail through
bf16 in place. Without that write-back the invariant would have held for
448 of 512 dims and the on/off gate would have gone red for the wrong
reason. Read-back stays bit-identical to the f32 cache.

The repack path narrows losslessly: its input is already-roundtripped
cache data, so those rope dims are exactly bf16-representable and the
second encode cannot drift -- the same property attn_pack_exact_e8_dev
buys for the nope dims. That matters because we re-encode on evict,
restore and fork, where vLLM's write-once cache never does.

Readers updated in step: both scalar/lane8 packed dots, both float4 smem
staging sites in pulsar_cuda_attention.cu, and the fp16 tensor-core
kernel's own rope branch in pulsar_cuda_attn_f16.cu, which transcribes the
tail rather than sharing attn_comp_pack_ld. The two gates that synthesise
packed rows by hand now emit bf16; attn_f16_banked_test's constant is
exact in e4m3 (3 mantissa bits) hence exact in bf16 (7), so its isolation
assertion is unchanged.

PULSAR_SESSION_PAYLOAD_VERSION 2 -> 3: a v2 disk payload's comp rows sit on
the old 712 B stride, so it must be refused rather than reinterpreted.

NOT YET BUILT OR MEASURED -- the dev box has no nvcc. The KL A/B against
the f32-rope arm and the predicted decode gain at depth are both still owed
before this is trustworthy.

+++ src/cuda/pulsar_cuda_attention.cu
@@ -40,9 +40,11 @@ __device__ static float pulsar_attn_fp8_kv_dequant(uint8_t byte, float scale) {
 
 /* comp_kv_pack: per-call flag — the comp cache operand is PULSAR_ATTN_PACK rows
  * (see PULSAR_ATTN_PACK_* in pulsar_cuda_internal.h): [n_nope e4m3][n_nope/64 E8M0]
- * [pad][n_rot f32 rope].  Nope dims decode e4m3_value * 2^(e8-127) — exactly
- * the fp8_kv_quantize roundtrip value the f32 cache holds — and rope dims read
- * f32 directly, so scores/outputs are bit-identical to the f32 comp cache.
+ * [pad][n_rot bf16 rope].  Nope dims decode e4m3_value * 2^(e8-127) — exactly
+ * the fp8_kv_quantize roundtrip value the f32 cache holds — and rope dims widen
+ * from bf16, which is likewise exactly what the f32 cache holds since the pack
+ * store roundtrips the rope tail in place, so scores/outputs are bit-identical
+ * to the f32 comp cache.
  * The 2^(e8-127) scale is built as a float exponent; the pack amax floor
  * (1e-4 -> e8 >= 105) rules out byte 0. */
 /* attn_pack_e4m3 / attn_comp_pack_ld now live in pulsar_cuda_internal.h so the
@@ -52,7 +54,7 @@ __device__ static float pulsar_attn_fp8_kv_dequant(uint8_t byte, float scale) {
 
 /* Packed-row dot walked d = 0..head_dim-1 by one thread: per-64-block scale
  * hoisted, e4m3 bytes fetched four at a time as one uint32 (rows are 4-byte
- * aligned: 712-byte stride).  Accumulation order is identical to the scalar
+ * aligned: 584-byte stride).  Accumulation order is identical to the scalar
  * d-ascending loop, so the result is bit-identical. */
 __device__ static inline float attn_pack_dot_full(const float *qh, const float *comp_kv, uint64_t row, uint32_t head_dim, float dot) {
     const uint32_t n_nope = head_dim - PULSAR_ATTN_PACK_NROT;
@@ -70,8 +72,8 @@ __device__ static inline float attn_pack_dot_full(const float *qh, const float *
             dot += qh[d + 3u] * attn_pack_e4m3(w >> 24, scale);
         }
     }
-    const float *rope = (const float *)(pr + n_nope + PULSAR_ATTN_PACK_SCALES_PAD(head_dim));
-    for (uint32_t d = 0; d < PULSAR_ATTN_PACK_NROT; d++) dot += qh[n_nope + d] * rope[d];
+    const __nv_bfloat16 *rope = (const __nv_bfloat16 *)(pr + n_nope + PULSAR_ATTN_PACK_SCALES_PAD(head_dim));
+    for (uint32_t d = 0; d < PULSAR_ATTN_PACK_NROT; d++) dot += qh[n_nope + d] * __bfloat162float(rope[d]);
     return dot;
 }
 
@@ -88,8 +90,8 @@ __device__ static inline float attn_pack_dot_lane8(const float *qh, const float 
             dot += qh[d] * attn_pack_e4m3(pr[d], scale);
         }
     }
-    const float *rope = (const float *)(pr + n_nope + PULSAR_ATTN_PACK_SCALES_PAD(head_dim));
-    for (uint32_t d = n_nope + lane; d < head_dim; d += 8u) dot += qh[d] * rope[d - n_nope];
+    const __nv_bfloat16 *rope = (const __nv_bfloat16 *)(pr + n_nope + PULSAR_ATTN_PACK_SCALES_PAD(head_dim));
+    for (uint32_t d = n_nope + lane; d < head_dim; d += 8u) dot += qh[d] * __bfloat162float(rope[d - n_nope]);
     return dot;
 }
 
@@ -1252,11 +1254,11 @@ __global__ PULSAR_ATTN_LB static void attention_indexed_mixed_heads8_online_kern
                     v.z = attn_pack_e4m3((w >> 16) & 0xffu, scale);
                     v.w = attn_pack_e4m3(w >> 24, scale);
                 } else {
-                    const float *rope = (const float *)(pr + n_nope + PULSAR_ATTN_PACK_SCALES_PAD(head_dim));
-                    v.x = rope[base - n_nope + 0u];
-                    v.y = rope[base - n_nope + 1u];
-                    v.z = rope[base - n_nope + 2u];
-                    v.w = rope[base - n_nope + 3u];
+                    const __nv_bfloat16 *rope = (const __nv_bfloat16 *)(pr + n_nope + PULSAR_ATTN_PACK_SCALES_PAD(head_dim));
+                    v.x = __bfloat162float(rope[base - n_nope + 0u]);
+                    v.y = __bfloat162float(rope[base - n_nope + 1u]);
+                    v.z = __bfloat162float(rope[base - n_nope + 2u]);
+                    v.w = __bfloat162float(rope[base - n_nope + 3u]);
                 }
                 kv_shared[off] = v;
             } else {
@@ -1883,11 +1885,11 @@ __global__ static void attention_decode_mixed_heads8_online_kernel(
                     v.z = attn_pack_e4m3((w >> 16) & 0xffu, scale);
                     v.w = attn_pack_e4m3(w >> 24, scale);
                 } else {
-                    const float *rope = (const float *)(pr + n_nope + PULSAR_ATTN_PACK_SCALES_PAD(head_dim));
-                    v.x = rope[base - n_nope + 0u];
-                    v.y = rope[base - n_nope + 1u];
-                    v.z = rope[base - n_nope + 2u];
-                    v.w = rope[base - n_nope + 3u];
+                    const __nv_bfloat16 *rope = (const __nv_bfloat16 *)(pr + n_nope + PULSAR_ATTN_PACK_SCALES_PAD(head_dim));
+                    v.x = __bfloat162float(rope[base - n_nope + 0u]);
+                    v.y = __bfloat162float(rope[base - n_nope + 1u]);
+                    v.z = __bfloat162float(rope[base - n_nope + 2u]);
+                    v.w = __bfloat162float(rope[base - n_nope + 3u]);
                 }
                 kv_shared[off] = v;
             } else {
